requirements.txt
allure-pytest==2.13.5
beautifulsoup4==4.12.3
pytest==8.3.1
pytest-xdist==3.6.1
rstr==3.2.2
selenium==4.24.0
lxml==5.2.2